Update test comments regarding unicode D800-DFFF
The D800-DBFF range are the high surrogates, and the DC00-DFFF range
are the low surrogates.  These are used in UTF-16 encoding to allow
one to represent a character beyond the Basic Multilingual Plane,
i.e. a value greater than 0xffff.

They are only valid when used in such pairs, and only when used in
a UTF-16 encoding. In all other contexts they are undefined, and
as such trying to decode them may trigger unexpected behaviour.