Fix go vet for Bullseye build
From Go1.12 onwards 'go tool vet' is deprecated, so need to switch to 'go vet'. This entails specifying the GOPATH, and in this repo, running the go generate commands explicitly as the switch to a more recent Go compiler and / or changes to dh_golang mean it no longer implicitly runs go generate for go vet even when the DH_GOLANG_GO_GENERATE flag is set in the rules file.

One complaint found by go vet has been fixed.

Upped compat level to 12 (max supported by Buster / Deb10) as 9 is deprecated in Bullseye / Deb11.

Build tested for Deb10 and Deb11