XPATH: Upon a lexer error, return ERR token not EOF
Jira: VRVDR-54901

If the lexer detects an invalid input pattern, it has to signal this
to the parser.  If EOF is returned, then the parsers will assume that
all input has been consumed, but will often detect an error due to an
unexpected EOF.

However if the lexer error occurs at the end of the text, after an
otherwise valid expression, the EOF will appear valid, and hence
the parser will not detect an error.

So a new token (ERR) in introduced and returned by the lexer for such
cases.  None of the parsers shall match this token, and so wherever it
is seen, it will cause an error.