XPATH: Allow parser to complete and return status
Jira: VRVDR-54901

Yacc generated parser will return 0 for success or 1 for failure.

Success occurs when the explicit top level rule is matched followed
by the end of input (EOF). This happens when that rule is allowed
to return to an implicit higher level rule similar to:

implicit: top EOF { return 0 }
        | { return 1 }
        ;

As such success is only returned if all input is consumed and parsed.

Any errors in the parse, including unmatched tokens will then result
in an error being returned.

This is sufficient to handle the trailing ')' and similar cases
currently caught by the post parse check for if all lexer text has
been consumed. e.g. text of "1)"

This would also handle empty input text (i.e. "") and all white space
input (e.g. " ") as these will be treated as an unexpected EOF,
and hence a parse error.

However that post parse check is still necessary in order to cope with
some invalid lexer tokens at the end of a valid parse,
e.g. "1 ! ".  That is because such lexer errors return an EOF token.