XPATH: Make the lexer reject invalid UTF-8 input
Jira: VRVDR-54901

Invalid UTF-8 encoded text should not simply be skipped as if
it was not present, in general such behaviour can lead to
latent security bugs.  Instead treat this as any other form
of lexer error, and signal as such to the parser, aborting
the parse.

Invalid UTF-8 during a NextNonWhitespaceStringIs() lookahead
is ignored, while forcing a failure from that lookahead.
The invalid text will eventually be detected when the caller
subsequently asks for more tokens; or possibly the parse will
fail early due some other error.  Either case is acceptable.

Need to update low level lexer tests now to prove the skip whitespace
case, and the explicit utf-8 error case