XPATH: Eliminate xpath.CommonSymType
Jira: VRVDR-54951

There is no need for this as the interface variable mechanism allows
for a token value to be represented as a form of typed 'union'.

When that is then combined with the multiple return values, we end up
with a reduction in code, and a simpler interface.  As a bonus it is
fractionally fasters, as it avoids copying to the unused fields in
the yacc 'union'.