Cache leafref allowed values.
When we have 1000 instances of a leafref node, all referencing another node that has 1000 instances, there's an awful lot of redundant regeneration of allowed values. When there is no predicate in the leafref path, we can simply cache the allowed values from the first time we get them (lookup on node path), and then reuse them the next 999 times.

Verified on a J2 (where config previously taking 40s to validate now takes 1.5s), and with unit tests in configd (separate PR).

Fixes VRVDR-55035