Fix over zealous leafref optimisation
Fixes VRVDR-56180.

Where a leafref uses a relative path and is wholly within a higher level list, values cached may in fact be for the wrong list element. Optimisation can't be used as-is for such paths so an additional exception is being added.  Unit tests in the accompanying configd PR verify the fix.

Note that VRVDR-54991 has been reopened to track the requirement for re-optimising if the J2 / GPC code ships.